Deepfake-Detection-System/deepfake_api/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
opencv-python
python-multipart
python-dotenv